requirements.txt
requests
python-dotenv
streamlit
fastapi
pydantic
apache-airflow-providers-docker
apache-airflow-providers-google
pandas
pyarrow
dbt-bigquery